feat: update API to match other proto encoding APIs
* Use MarshalOptions to enable inclusion of future options

* Top-level Marshal function uses default MarshalOptions

* Schema inference also uses MarshalOptions (to ensure compatibility
  between inferred schema and marshalled value).
