+++ encoding/protobq/doc.go
@@ -1,0 +1,3 @@
+// Package protobq marshals and unmarshals protocol buffer messages as BigQuery
+// format.
+package protobq
