+++ encoding/protobq/encode_test.go
@@ -9,10 +9,64 @@ import (
 	"gotest.tools/v3/assert"
 )
 
-func TestSchema(t *testing.T) {
+func TestMarshalOptions_Marshal(t *testing.T) {
 	for _, tt := range []struct {
 		name     string
 		msg      proto.Message
+		opt      MarshalOptions
+		expected map[string]bigquery.Value
+	}{
+		{
+			name: "library.Book",
+			msg: &library.Book{
+				Name:   "name",
+				Author: "author",
+				Title:  "title",
+				Read:   true,
+			},
+			expected: map[string]bigquery.Value{
+				"name":   "name",
+				"author": "author",
+				"title":  "title",
+				"read":   true,
+			},
+		},
+		{
+			name: "library.UpdateBookRequest",
+			msg: &library.UpdateBookRequest{
+				Name: "name",
+				Book: &library.Book{
+					Name:   "name",
+					Author: "author",
+					Title:  "title",
+					Read:   true,
+				},
+			},
+			expected: map[string]bigquery.Value{
+				"name": "name",
+				"book": map[string]bigquery.Value{
+					"name":   "name",
+					"author": "author",
+					"title":  "title",
+					"read":   true,
+				},
+			},
+		},
+	} {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			actual, err := tt.opt.Marshal(tt.msg)
+			assert.NilError(t, err)
+			assert.DeepEqual(t, tt.expected, actual)
+		})
+	}
+}
+
+func TestMarshalOptions_InferSchema(t *testing.T) {
+	for _, tt := range []struct {
+		name     string
+		msg      proto.Message
+		opt      MarshalOptions
 		expected bigquery.Schema
 	}{
 		{
@@ -45,7 +99,7 @@ func TestSchema(t *testing.T) {
 	} {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
-			actual := Schema(tt.msg)
+			actual := tt.opt.InferSchema(tt.msg)
 			assert.DeepEqual(t, tt.expected, actual)
 		})
 	}
