+++ encoding/protobq/messagesaver.go
@@ -5,15 +5,34 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
+// MessageSaver implements bigquery.ValueSaver for a proto.Message.
+// The message is converted to a BigQuery row using the provided MarshalOptions.
 type MessageSaver struct {
-	Message  proto.Message
+	// Options to use for marshaling the Message.
+	Options MarshalOptions
+
+	// InsertID governs the best-effort deduplication feature of
+	// BigQuery streaming inserts.
+	//
+	// If the InsertID is empty, a random InsertID will be generated by
+	// this library to facilitate deduplication.
+	//
+	// If the InsertID is set to the sentinel value bigquery.NoDedupeID, an InsertID
+	// is not sent.
+	//
+	// For all other non-empty values, BigQuery will use the provided
+	// value for best-effort deduplication.
 	InsertID string
+
+	// Message to save.
+	Message proto.Message
 }
 
 var _ bigquery.ValueSaver = &MessageSaver{}
 
+// Save implements bigquery.ValueSaver.
 func (m *MessageSaver) Save() (map[string]bigquery.Value, string, error) {
-	row, err := Marshal(m.Message)
+	row, err := m.Options.Marshal(m.Message)
 	if err != nil {
 		return nil, "", err
 	}
